% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_num_payments.R
\name{loan_num_payments}
\alias{loan_num_payments}
\title{Calculate the Total Number of Payments for a Loan}
\usage{
loan_num_payments(loan_amount, interest_rate, monthly_payment)
}
\arguments{
\item{loan_amount}{total amount of loan}

\item{interest_rate}{iterest rate of loan}

\item{monthly_payment}{monthly loan payments}
}
\value{
returns amount of monthly installments for the loan balance
}
\description{
Calculate the Total Number of Payments for a Loan
}
\examples{
loan_num_payments(loan_amount= 25000, interest_rate = 2, monthly_payment = 234)
}
