% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_payment.R
\name{loan_payment}
\alias{loan_payment}
\title{Calculate Monthly Loan Payments}
\usage{
loan_payment(loan_amount, interest_rate, loan_term)
}
\arguments{
\item{loan_amount}{amount of the loan}

\item{interest_rate}{rate for interest on loan}

\item{loan_term}{amount of years for loan}
}
\value{
total monthly payments on loan
}
\description{
Calculate Monthly Loan Payments
}
\examples{
loan_payment(loan_amount = 20000, interest_rate = 5, loan_term = 24)
}
