% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_interest.R
\name{total_interest}
\alias{total_interest}
\title{Calculate the Total Interest on a Loan}
\usage{
total_interest(loan_amount, interest_rate, loan_term)
}
\arguments{
\item{loan_amount}{Total loan amount}

\item{interest_rate}{Total interest rate for loan amount}

\item{loan_term}{Length of loan payments in years}
}
\value{
returns total interest on a loan
}
\description{
Calculate the Total Interest on a Loan
}
\examples{
total_interest(loan_amount = 500 , interest_rate = 3, loan_term = 3)
}
